Add files via upload
Starting PRAW and Getting Data from Reddit:-
Installing and importing the required libraries—PRAW, pandas, numpy, and regular expressions (re)—is the first step in the code.A user agent and some client credentials are used to create a Reddit instance. Next, the code uses the PRAW library to retrieve data from the "bitcoin" subreddit.In the "hot" area of the subreddit, pertinent details are printed for each entry, including the title, ID, author, creation time, score, upvote ratio, and URL.The script is programmed to gather information up until it comes across a submission (post), at which point it exits the loop.

Data Cleaning and Preparation:-
The information is saved in a DataFrame named bitcoin_df after it has been retrieved.The titles of the Reddit submissions are originally displayed in a single column of the DataFrame.A function named cleanTxt is used to clean up titles by eliminating textual elements such as hashtags, retweet tags, hyperlinks, and colons.To remove Unicode characters and emojis from the titles, using the remove_emoji function.The DataFrame is filled with the cleaned titles again.

Sentiment Analysis: TextBlob is used to do sentiment analysis on the cleaned titles.To determine the subjectivity and polarity of each title, two functions are defined: getSubjectivity and getPolarity.Subjectivity and polarity, the sentiment analysis results, are added to the DataFrame as additional columns.

Sentiment Categorization: Using the polarity score of each title, a new column named "Insight" is made to classify the sentiment as "Negative," "Neutral," or "Positive".Based on the polarity score, sentiment categories are assigned using a function named getInsight.

Data Visualization:- Data visualization is accomplished by importing the Seaborn and matplotlib libraries.The sentiment scores are visualized using a bar plot. Sentiment categories ("Negative," "Neutral," or "Positive") are represented by the x-axis, while the frequency of each sentiment type is displayed on the y-axis.To see the words that appear most frequently in the titles, a word cloud is created. Every word in the word cloud has a size and color that correspond to its importance and frequency, respectively.